pagination in signals, correcting controllers and services everywhere

+++ THManager/THManager/Models/AccountTradeHistory.cs
@@ -11,15 +11,15 @@ namespace THManager.Models
         public DateTime Time { get; set; }
         public string Side { get; set; }
         public string Pair { get; set; }
-        public double Price { get; set; }
-        public double Quantity { get; set; }
-        public double DesiredQuantity { get; set; }
-        public double DollarQuantity { get; set; }
-        public double DesiredDollarQuantity { get; set; }
-        public double Profit { get; set; }
-        public double DesiredProfit { get; set; }
-        public double PercentProfit { get; set; }
-        public double DesiredPercentProfit { get; set; }
+        public decimal Price { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal DesiredQuantity { get; set; }
+        public decimal DollarQuantity { get; set; }
+        public decimal DesiredDollarQuantity { get; set; }
+        public decimal Profit { get; set; }
+        public decimal DesiredProfit { get; set; }
+        public decimal PercentProfit { get; set; }
+        public decimal DesiredPercentProfit { get; set; }
         public string SignalStr { get; set; }
     }
 }
